requirements.txt
asgiref==3.5.2
certifi==2022.6.15
Django==4.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-recaptcha==3.0.0
django-tinymce==3.5.0
fontawesomefree==6.1.2
Pillow==9.2.0
six==1.16.0
sqlparse==0.4.2
